requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.23.0
Pillow>=9.0.0
scipy>=1.9.0
tqdm>=4.64.0
einops>=0.6.0
timm>=0.6.0
opencv-python>=4.6.0
tensorboard>=2.11.0
